progressBar added, request cancellation code done, onPreExecute onPostExecute downloadInBackground functionality achieved
