--- app/src/main/java/com/example/customasynctask/CustomHandler.kt
@@ -1,17 +1,0 @@
-package com.example.customasynctask
-
-import android.os.Handler
-import android.os.Message
-
-class CustomHandler(private val response: DataReceiver) : Handler(){
-
-    override fun handleMessage(msg: Message) {
-        super.handleMessage(msg)
-        response.Response(msg)
-    }
-
-    interface DataReceiver{
-
-        fun Response(message: Message)
-    }
-}
